+++ applications/amathea/src/features/dataviews/NewDataview.tsx
@@ -3,12 +3,12 @@ import { useSelector } from 'react-redux'
 import Select, { SelectOption } from '@globalfishingwatch/ui-components/dist/select'
 import Button from '@globalfishingwatch/ui-components/dist/button'
 import ColorBar, { ColorBarIds } from '@globalfishingwatch/ui-components/dist/color-bar'
-import { DATASET_SOURCE_OPTIONS, FLAG_FILTERS } from 'data/data'
+import { DATASET_SOURCE_OPTIONS } from 'data/data'
 import { useModalConnect } from 'features/modal/modal.hooks'
 import { useCurrentWorkspaceConnect, useWorkspacesAPI } from 'features/workspaces/workspaces.hook'
 import styles from './NewDataview.module.css'
 import { selectDatasetOptionsBySource } from './dataviews.selectors'
-import { useDraftDataviewConnect, useDataviewsAPI, useDataviewsConnect } from './dataviews.hook'
+import { useDraftDataviewConnect, useDataviewsAPI } from './dataviews.hook'
 import { DataviewDraftDataset } from './dataviews.slice'
 
 function NewDataview(): React.ReactElement {
@@ -18,20 +18,15 @@ function NewDataview(): React.ReactElement {
   const { updateWorkspace } = useWorkspacesAPI()
   const { draftDataview, setDraftDataview, resetDraftDataview } = useDraftDataviewConnect()
   const { createDataview } = useDataviewsAPI()
-  const { dataviewsList } = useDataviewsConnect()
   const { source, dataset } = draftDataview || {}
   const datasetsOptions = useSelector(selectDatasetOptionsBySource)
+
   const onSelect = (property: string, option: SelectOption | DataviewDraftDataset | number) => {
     setDraftDataview({ [property]: option })
   }
-  const onDatasetSelect = (dataset: SelectOption) => {
-    if (draftDataview?.source?.id === 'gfw') {
-      const dataview = dataviewsList.find((dataview) =>
-        dataview.datasets?.find((d) => d.id === dataset.id)
-      )
-      if (dataview) {
-        onSelect('id', dataview.id)
-      }
+  const onDatasetSelect = (dataset: any) => {
+    if (dataset.dataviewId) {
+      onSelect('id', dataset.dataviewId)
     }
     onSelect('dataset', dataset)
   }
@@ -69,8 +64,8 @@ function NewDataview(): React.ReactElement {
     }
   }
 
-  const isFishingEffortLayer = dataset?.id === 'dgg_fishing_galapagos'
-  const selectedFlagFilter = FLAG_FILTERS.find((flag) => flag.id === draftDataview?.flagFilter)
+  // const isFishingEffortLayer = dataset?.id === 'dgg_fishing_galapagos'
+  // const selectedFlagFilter = FLAG_FILTERS.find((flag) => flag.id === draftDataview?.flagFilter)
   return (
     <div className={styles.container}>
       <h1 className="screen-reader-only">{draftDataview?.id ? 'Update Dataset' : 'New Dataset'}</h1>
@@ -103,7 +98,7 @@ function NewDataview(): React.ReactElement {
           />
         </div>
       )}
-      {isFishingEffortLayer && (
+      {/* {isFishingEffortLayer && (
         <Select
           label="Filter by flag"
           options={FLAG_FILTERS}
@@ -113,7 +108,7 @@ function NewDataview(): React.ReactElement {
           onRemove={() => onCleanClick('flagFilter')}
           onCleanClick={() => onCleanClick('flagFilter')}
         ></Select>
-      )}
+      )} */}
       <Button onClick={onCreateClick} className={styles.saveBtn} loading={loading}>
         Confirm
       </Button>