+++ packages/layer-composer/src/generators/types.ts
@@ -157,6 +157,11 @@ export interface RulersGeneratorConfig extends GeneratorConfig {
   data: Ruler[]
 }
 
+export interface HeatmapGeneratorLegend {
+  label: string
+  unit: string
+}
+
 export interface HeatmapGeneratorConfig extends GeneratorConfig {
   type: Type.Heatmap
   // Types needed but already in GlobalGeneratorConfig
@@ -164,6 +169,7 @@ export interface HeatmapGeneratorConfig extends GeneratorConfig {
   // end: string
   // zoom: number
   maxZoom?: number
+  steps?: number[]
   tilesUrl: string
   statsUrl?: string
   fetchStats?: boolean
@@ -172,6 +178,7 @@ export interface HeatmapGeneratorConfig extends GeneratorConfig {
   colorRamp?: ColorRampsIds
   serverSideFilter?: string
   updateColorRampOnTimeChange?: boolean
+  legend?: HeatmapGeneratorLegend
 }
 
 export interface HeatmapAnimatedGeneratorConfig extends GeneratorConfig {
@@ -252,5 +259,7 @@ export type ColorRampsIds =
   | 'yellow'
   | 'green'
   | 'orange'
+
+export type BivariateColorRampsIds = 'bivariate'
 
 export type CombinationMode = 'add' | 'compare' | 'bivariate' | 'literal'