+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [e6f5aa7]
+  - @globalfishingwatch/react-hooks@2.8.0
+  - @globalfishingwatch/layer-composer@2.8.1
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [a566ff6]
+  - @globalfishingwatch/layer-composer@2.8.0
+  - @globalfishingwatch/react-hooks@2.7.0
+
 ## 0.2.3
 
 ### Patch Changes