+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @globalfishingwatch/react-hooks
 
+## 2.8.0
+
+### Minor Changes
+
+- e6f5aa7: Update enpdoints ids
+
+### Patch Changes
+
+- Updated dependencies [e6f5aa7]
+  - @globalfishingwatch/layer-composer@2.8.1
+
+## 2.7.0
+
+### Minor Changes
+
+- a566ff6: Layer component
+
+### Patch Changes
+
+- Updated dependencies [a566ff6]
+  - @globalfishingwatch/layer-composer@2.8.0
+
 ## 2.6.1
 
 ### Patch Changes